# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Machine à Café Connectée",
        "description": "API pour simuler une machine à café connectée",
        "contact": {
            "email": "contact@cafe-connecte.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/v1/process/{order}/start": {
            "post": {
                "tags": [
                    "Process"
                ],
                "summary": "Démarre le processus de préparation",
                "operationId": "5c75d967dd16323d207275df5d287124",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Processus démarré"
                    },
                    "400": {
                        "description": "Commande déjà en cours"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                }
            }
        },
        "/api/v1/process/{order}/progress": {
            "get": {
                "tags": [
                    "Process"
                ],
                "summary": "Récupère la progression du processus",
                "operationId": "0f82bffdaf51db6659548c5fc31727d0",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Progression du processus"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                }
            }
        },
...
        "/api/v1/queue/validate-slot": {
            "post": {
                "tags": [
                    "Queue"
                ],
                "summary": "Valide si un créneau est disponible",
                "operationId": "786b9b92aa3e58aa20e8395d5cab42a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pickup_time"
                                ],
                                "properties": {
                                    "pickup_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-05-11 15:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Résultat de la validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_available": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "requested_time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "earliest_available_time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Créneau disponible"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}